+++ video-editor/src/main/java/com/video/trimmer/interfaces/OnProgressVideoListener.kt
@@ -1,5 +1,5 @@
 package com.video.trimmer.interfaces
 
 interface OnProgressVideoListener {
-    fun updateProgress(time: Int, max: Int, scale: Float)
+    fun updateProgress(time: Float, max: Float, scale: Float)
 }
